requirements.txt
httpx>=0.27.0
PyJWT>=2.8.0
pytest>=8.0.0
